+++ CHANGES.rst
@@ -5,6 +5,7 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
 
 - Set page_size for events requests to 5000.
+- Round start & end timestamps on timeseries export URL.
 
 
 Release 3.0.4 (2016-4-5)